% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_func.r
\name{anovaben}
\alias{anovaben}
\title{Computes ANOVA test}
\usage{
anovaben(objectives, data, ffd, obj_results, alpha)
}
\arguments{
\item{objectives}{User selected objectives.}

\item{data}{Peak information (all times, heights, etc.).}

\item{ffd}{Fractional factorial design}

\item{obj_results}{Calculated objectives.}

\item{alpha}{Alpha value.}
}
\value{
list containing the results of ANOVA on the data from the FFD
(including p-values and which parameters are significant), the suggested
changes to add a level for a BBD, suggested values for the BBD, row indices
for the results, and the data frame of the FFD with objective results included
}
\description{
Runs an ANOVA test for any variables that were selected by the user.
}
