% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_func.r
\name{area}
\alias{area}
\title{Computes areas of peaks}
\usage{
area(data, hts, tms)
}
\arguments{
\item{data}{dataframe of time versus intensity.}

\item{hts}{heights of peaks in chronological order.}

\item{tms}{time locations of peaks in chronological order.}
}
\value{
area for each peak
}
\description{
Computes the areas beneath the specified maxima.
}
