% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blc_func.r
\name{blc}
\alias{blc}
\title{Baseline correction}
\usage{
blc(frame, noise = 10^5, subtract = NULL)
}
\arguments{
\item{frame}{data frame of time versus intensity.}

\item{noise}{strength of the baseline subtraction.}

\item{subtract}{how much intensity to initially subtract.}
}
\value{
data frame of of data that has been baseline corrected
}
\description{
Generates the baseline of the spectra, interpolates between points, and subtracts from the intensity to generate corrected baseline.
}
