% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_func.r
\name{gauss}
\alias{gauss}
\title{Generate a Gaussian fit}
\usage{
gauss(a, b, c, t)
}
\arguments{
\item{a}{amplitute.}

\item{b}{phase shift.}

\item{c}{standard deviation.}

\item{t}{time at which to fit.}
}
\value{
value of Gaussian function
}
\description{
Generate a Gaussian fit with given parameters.
}
