% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodOpt.R
\name{MethodOpt}
\alias{MethodOpt}
\title{Method optimization GUI}
\usage{
MethodOpt()
}
\value{
{No return value, opens and runs the GUI}
}
\description{
`MethodOpt()` runs the GUI.
}
\details{
No arguments are needed to initiate the GUI. `MethodOpt()` is split into three main tabs.
}
\section{Fractional factorial design}{

The first step in method optimization is to build an experimental design. Hence,
the first tab of the GUI is dedicated to designing a fractional factorial experimental design.
Parameters are input with their corresponding low and high values. Pressing "Generate FFD" will
yield the experimental design. The user will run experiments according to each method.
}

\section{ANOVA}{

The second step in method optimization is to run an ANOVA test. This is carried out
under the "Analysis" tab. Raw experimental screening data is uploaded, and the spectra
can be viewed. In subtabs, the spectra peaks must be identified (either by an uploaded retention time
file or by a built-in identification algorithm), objectives must be selected, and the initial
experimental design must be uploaded; then the ANOVA test may be run.
Statistically significant parameters are indicated.
}

\section{Box-Behnken design}{

The third step is to generate a three-level Box-Behnken experimental design with the
significant parameters. Low, middle, and high values are input with their corresponding
parameters. The design can be generated by pressing "Generate BBD."
}

\section{Optimization}{

The final step is to run the optimization with the results of the Box-Behnken design.
Similarly to the second step, raw data is uploaded in the "Analysis" tab. Spectra peaks
must be identified, objectives
must be selected, and the experimental design must be uploaded; then the optimal
values can be calculated.
}

\examples{
# Please see the vignette for the MethodOpt package for a full example of how
# to use the GUI launched by MethodOpt::MethodOpt().

}
