% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_func.r
\name{opt}
\alias{opt}
\title{Optimization algorithm}
\usage{
opt(objectives, bbd, results, lim_fac, valid_range_data)
}
\arguments{
\item{objectives}{objectives input by user.}

\item{bbd}{Box-Behnken design.}

\item{results}{objective results.}

\item{lim_fac}{limiting factors.}

\item{valid_range_data}{ranges of validity corresponding to the limiting factors.}
}
\value{
a list containing the parameters which cannot be set to the unbounded
solution, the maximum value of the objectives, and the unbounded parameter
solutions
}
\description{
Calculate the optimal parameter values for given objectives.
}
