% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_rts_func.R
\name{peaks_rts}
\alias{peaks_rts}
\title{peak calculations with retention times}
\usage{
peaks_rts(raw_data, retention_times, rt_index, bl_noise = 0)
}
\arguments{
\item{raw_data}{time versus intensity dataframe.}

\item{retention_times}{retention time file.}

\item{rt_index}{which method to evaluate for.}

\item{bl_noise}{constant level at which response should be considered as noise.}
}
\value{
list containing the peaks' heights, times where they occur, the full
data TIC, full time index, and the peaks' widths
}
\description{
Locate the maxima of the peaks corresponding to the retention times.
}
