% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widths_func.r
\name{widths}
\alias{widths}
\title{Calculate widths}
\usage{
widths(data, hts, tms)
}
\arguments{
\item{data}{time versus intensity dataframe.}

\item{hts}{heights of peaks.}

\item{tms}{times of peaks.}
}
\value{
width for each peak
}
\description{
Calculate the widths (standard deviation) for each identified peak.
}
